Initial commit
Very, very simple configuration, it is a working progress that will be
in slow progress, but at least is something.

I'd like to see better documentation for fnordmetric, but I guess that
I'll figure it out by myself and then, as a good citizen, I should
send a pull request with better documentation.